add review feedback

+++ nexus-client/src/lib.rs
@@ -17,7 +17,7 @@ impl omicron_common::api::external::ClientError for types::Error {
 
 impl From<omicron_common::api::external::ByteCount> for types::ByteCount {
     fn from(s: omicron_common::api::external::ByteCount) -> Self {
-        Self(s.to_bytes() as i64)
+        Self(s.to_bytes())
     }
 }
